requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.1.0
av==13.1.0
benchmarl==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.0.4
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
distlib==0.3.9
dm-tree==0.1.9
docker-pycreds==0.4.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
gym==0.23.1
gym-notices==0.0.8
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
lz4==4.4.3
MarkupSafe==3.0.2
matplotlib==3.10.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numpy>=1.24,<2.0
omegaconf==2.3.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pip==25.0
platformdirs==4.3.6
proglog==0.1.10
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
pydantic==1.10.21
pydantic_core==2.20.1
pyglet==1.5.27
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyVirtualDisplay==3.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.5.3
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentry-sdk==2.21.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
tensordict==0.7.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.6.0
torch_cluster==1.6.3
torch-geometric==2.6.1
torchrl==0.7.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
virtualenv==20.29.2
vmas==1.5.0
wadler_lindig==0.1.4
wandb==0.17.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
